Fix bug in functions with early escapes
For `if` and other conditions that have a block statement, and
potentially have a `return` statement in one of the branches, we need to
add a `console.groupEnd` right before `return` to reset the indentation.
Otherwise, if the function containing the `if` condition gets called
multiple times, the indentation keeps on increasing.